Tests are working, regrouped them into single files, answered some PR… (#41)
* Tests are working, regrouped them into single files, answered some PR issues

* adjusted error checking in frame to remove redundancies.

* PR issues resolvement

* Removed redundant error checking from frame.go

* removed error checking on writeOpCode

* removed error checking from readShort

* removed error checking from readShortBytes
