+++ frame/response/error.go
@@ -18,16 +18,16 @@ func ParseError(b *frame.Buffer) Error {
 	}
 }
 
-// UnavailableErr spec: https://github.com/apache/cassandra/blob/trunk/doc/native_protocol_v4.spec#L1060
-type UnavailableErr struct {
+// UnavailableError spec: https://github.com/apache/cassandra/blob/trunk/doc/native_protocol_v4.spec#L1060
+type UnavailableError struct {
 	Error
 	Consistency frame.Consistency
 	Required    frame.Int
 	Alive       frame.Int
 }
 
-func ParseUnavailable(b *frame.Buffer) UnavailableErr {
-	return UnavailableErr{
+func ParseUnavailableError(b *frame.Buffer) UnavailableError {
+	return UnavailableError{
 		Error: Error{
 			Code:    b.ReadErrorCode(),
 			Message: b.ReadString(),
@@ -38,8 +38,8 @@ func ParseUnavailable(b *frame.Buffer) UnavailableErr {
 	}
 }
 
-// WriteTimeoutErr spec: https://github.com/apache/cassandra/blob/trunk/doc/native_protocol_v4.spec#L1076
-type WriteTimeoutErr struct {
+// WriteTimeoutError spec: https://github.com/apache/cassandra/blob/trunk/doc/native_protocol_v4.spec#L1076
+type WriteTimeoutError struct {
 	Error
 	Consistency frame.Consistency
 	Received    frame.Int
@@ -47,8 +47,8 @@ type WriteTimeoutErr struct {
 	WriteType   frame.WriteType
 }
 
-func ParseWriteTimeout(b *frame.Buffer) WriteTimeoutErr {
-	return WriteTimeoutErr{
+func ParseWriteTimeoutError(b *frame.Buffer) WriteTimeoutError {
+	return WriteTimeoutError{
 		Error: Error{
 			Code:    b.ReadErrorCode(),
 			Message: b.ReadString(),
@@ -60,8 +60,8 @@ func ParseWriteTimeout(b *frame.Buffer) WriteTimeoutErr {
 	}
 }
 
-// ReadTimeoutErr spec: https://github.com/apache/cassandra/blob/trunk/doc/native_protocol_v4.spec#L1108
-type ReadTimeoutErr struct {
+// ReadTimeoutError spec: https://github.com/apache/cassandra/blob/trunk/doc/native_protocol_v4.spec#L1108
+type ReadTimeoutError struct {
 	Error
 	Consistency frame.Consistency
 	Received    frame.Int
@@ -69,8 +69,8 @@ type ReadTimeoutErr struct {
 	DataPresent frame.Byte
 }
 
-func ParseReadTimeout(b *frame.Buffer) ReadTimeoutErr {
-	return ReadTimeoutErr{
+func ParseReadTimeoutError(b *frame.Buffer) ReadTimeoutError {
+	return ReadTimeoutError{
 		Error: Error{
 			Code:    b.ReadErrorCode(),
 			Message: b.ReadString(),
@@ -82,8 +82,8 @@ func ParseReadTimeout(b *frame.Buffer) ReadTimeoutErr {
 	}
 }
 
-// ReadFailureErr spec: https://github.com/apache/cassandra/blob/trunk/doc/native_protocol_v4.spec#L1125
-type ReadFailureErr struct {
+// ReadFailureError spec: https://github.com/apache/cassandra/blob/trunk/doc/native_protocol_v4.spec#L1125
+type ReadFailureError struct {
 	Error
 	Consistency frame.Consistency
 	Received    frame.Int
@@ -92,8 +92,8 @@ type ReadFailureErr struct {
 	DataPresent frame.Byte
 }
 
-func ParseReadFailure(b *frame.Buffer) ReadFailureErr {
-	return ReadFailureErr{
+func ParseReadFailureError(b *frame.Buffer) ReadFailureError {
+	return ReadFailureError{
 		Error: Error{
 			Code:    b.ReadErrorCode(),
 			Message: b.ReadString(),
@@ -106,16 +106,16 @@ func ParseReadFailure(b *frame.Buffer) ReadFailureErr {
 	}
 }
 
-// FuncFailureErr spec: https://github.com/apache/cassandra/blob/trunk/doc/native_protocol_v4.spec#L1140
-type FuncFailureErr struct {
+// FuncFailureError spec: https://github.com/apache/cassandra/blob/trunk/doc/native_protocol_v4.spec#L1140
+type FuncFailureError struct {
 	Error
 	Keyspace string
 	Function string
 	ArgTypes frame.StringList
 }
 
-func ParseFuncFailure(b *frame.Buffer) FuncFailureErr {
-	return FuncFailureErr{
+func ParseFuncFailureError(b *frame.Buffer) FuncFailureError {
+	return FuncFailureError{
 		Error: Error{
 			Code:    b.ReadErrorCode(),
 			Message: b.ReadString(),
@@ -126,8 +126,8 @@ func ParseFuncFailure(b *frame.Buffer) FuncFailureErr {
 	}
 }
 
-// WriteFailureErr spec: https://github.com/apache/cassandra/blob/trunk/doc/native_protocol_v4.spec#L1147
-type WriteFailureErr struct {
+// WriteFailureError spec: https://github.com/apache/cassandra/blob/trunk/doc/native_protocol_v4.spec#L1147
+type WriteFailureError struct {
 	Error
 	Consistency frame.Consistency
 	Received    frame.Int
@@ -136,8 +136,8 @@ type WriteFailureErr struct {
 	WriteType   frame.WriteType
 }
 
-func ParseWriteFailure(b *frame.Buffer) WriteFailureErr {
-	return WriteFailureErr{
+func ParseWriteFailureError(b *frame.Buffer) WriteFailureError {
+	return WriteFailureError{
 		Error: Error{
 			Code:    b.ReadErrorCode(),
 			Message: b.ReadString(),
@@ -150,15 +150,15 @@ func ParseWriteFailure(b *frame.Buffer) WriteFailureErr {
 	}
 }
 
-// AlreadyExistsErr spec: https://github.com/apache/cassandra/blob/trunk/doc/native_protocol_v4.spec#L1187
-type AlreadyExistsErr struct {
+// AlreadyExistsError spec: https://github.com/apache/cassandra/blob/trunk/doc/native_protocol_v4.spec#L1187
+type AlreadyExistsError struct {
 	Error
 	Keyspace string
 	Table    string
 }
 
-func ParseAlreadyExists(b *frame.Buffer) AlreadyExistsErr {
-	return AlreadyExistsErr{
+func ParseAlreadyExistsError(b *frame.Buffer) AlreadyExistsError {
+	return AlreadyExistsError{
 		Error: Error{
 			Code:    b.ReadErrorCode(),
 			Message: b.ReadString(),
@@ -168,14 +168,14 @@ func ParseAlreadyExists(b *frame.Buffer) AlreadyExistsErr {
 	}
 }
 
-// UnpreparedErr spec: https://github.com/apache/cassandra/blob/trunk/doc/native_protocol_v4.spec#L1197
-type UnpreparedErr struct {
+// UnpreparedError spec: https://github.com/apache/cassandra/blob/trunk/doc/native_protocol_v4.spec#L1197
+type UnpreparedError struct {
 	Error
-	UnknownID frame.Bytes
+	UnknownID frame.ShortBytes
 }
 
-func ParseUnprepared(b *frame.Buffer) UnpreparedErr {
-	return UnpreparedErr{
+func ParseUnpreparedError(b *frame.Buffer) UnpreparedError {
+	return UnpreparedError{
 		Error: Error{
 			Code:    b.ReadErrorCode(),
 			Message: b.ReadString(),
