+++ frame/response/supported_test.go
@@ -1,0 +1,3 @@
+package response
+
+// Nothing here... yet
