+++ go.mod
@@ -2,4 +2,4 @@ module scylla-go-driver
 
 go 1.17
 
-require github.com/google/go-cmp v0.5.6 // indirect
+require github.com/google/go-cmp v0.5.6
